Move datalayer sample to datalayer sample app
