+++ datalayer/sample/wear/src/main/java/com/google/android/horologist/datalayer/sample/Screen.kt
@@ -22,4 +22,6 @@ sealed class Screen(
     object MainScreen : Screen("mainScreen")
 
     object DataLayerScreen : Screen("dataLayerScreen")
+
+    object DataLayerNodes : Screen("nodes")
 }
