+++ datalayer/sample/wear/src/main/java/com/google/android/horologist/datalayer/sample/screens/nodes/SampleDataSerializer.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Android Open Source Project
+ * Copyright 2023 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.horologist.datalayer
+package com.google.android.horologist.datalayer.sample.screens.nodes
 
 import androidx.datastore.core.CorruptionException
 import androidx.datastore.core.Serializer
