+++ datalayer/sample/wear/src/main/java/com/google/android/horologist/datalayer/sample/screens/nodes/SampleDataService.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Android Open Source Project
+ * Copyright 2023 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package com.google.android.horologist.datalayer
+package com.google.android.horologist.datalayer.sample.screens.nodes
 
-import com.google.android.horologist.components.SampleApplication
 import com.google.android.horologist.data.WearDataLayerRegistry
 import com.google.android.horologist.data.WearDataService
+import com.google.android.horologist.datalayer.sample.SampleApplication
 
 class SampleDataService : WearDataService() {
     override lateinit var registry: WearDataLayerRegistry
