+++ datalayer/sample/wear/src/main/proto/sample.proto
