+++ sample/src/main/java/com/google/android/horologist/components/SampleApplication.kt
@@ -17,11 +17,6 @@
 package com.google.android.horologist.components
 
 import android.app.Application
-import com.google.android.horologist.data.ProtoDataStoreHelper.registerProtoDataListener
-import com.google.android.horologist.data.WearDataLayerRegistry
-import com.google.android.horologist.data.proto.SampleProto
-import com.google.android.horologist.data.store.ProtoDataListener
-import com.google.android.horologist.datalayer.SampleDataSerializer
 import com.google.android.horologist.networks.InMemoryStatusLogger
 import com.google.android.horologist.networks.data.DataRequestRepository
 import com.google.android.horologist.networks.status.NetworkRepository
@@ -38,25 +33,9 @@ class SampleApplication : Application() {
     lateinit var dataRequestRepository: DataRequestRepository
     lateinit var networkAwareCallFactory: Call.Factory
 
-    lateinit var registry: WearDataLayerRegistry
-
     override fun onCreate() {
         super.onCreate()
 
         SampleAppDI.inject(this)
-
-        registry = WearDataLayerRegistry.fromContext(this, coroutineScope).apply {
-            registerSerializer(SampleDataSerializer)
-
-            registerProtoDataListener(object : ProtoDataListener<SampleProto.Data> {
-                override fun dataAdded(nodeId: String, path: String, value: SampleProto.Data) {
-                    println("Data Added: $nodeId $path $value")
-                }
-
-                override fun dataDeleted(nodeId: String, path: String) {
-                    println("Data Deleted: $nodeId $path")
-                }
-            })
-        }
     }
 }
