+++ sample/src/main/java/com/google/android/horologist/sample/MenuChips.kt
@@ -210,19 +210,6 @@ fun NetworkChip(
 }
 
 @Composable
-fun DataLayerNodesChip(
-    navigateToRoute: () -> Unit,
-) {
-    SampleChip(
-        onClick = { navigateToRoute() },
-        label = "Data Layer Nodes",
-        content = {
-            Text(text = "Nodes")
-        },
-    )
-}
-
-@Composable
 fun PagingChip(
     navigateToRoute: () -> Unit,
 ) {
