+++ sample/src/main/java/com/google/android/horologist/sample/MenuScreen.kt
@@ -55,9 +55,6 @@ fun MenuScreen(
             NetworkChip { navigateToRoute(Screen.Network.route) }
         }
         item {
-            DataLayerNodesChip { navigateToRoute(Screen.DataLayerNodes.route) }
-        }
-        item {
             FillMaxRectangleChip(navigateToRoute)
         }
         item {
